Add search component in application
A new search component has been introduced to the application. This component, includes an input field for searching and a table which lists the 'loopNumber', 'gender', 'coat', 'birthDate', and 'pasture' properties of the 'Bovine' object. The component also contains a 'filter' function which filters the table based on the typed input.