Update navbar, header and footer components
The navbar, header and footer components have been updated with new functionalities. Angular Material elements have been added to the navbar, along with routing links for navigation. Correspondingly, import statements for "MatToolbar" and "MatIcon" are added to the AppModule. The header and footer components are also updated with more meaningful content and stylistic changes.