Update pastures component and inject additional features
The pastures component has been revamped with better styling and increased functionality. Additional changes include the integration of detailed pasture cards in a grid for different screen sizes, inclusion of a new "Vente" button in the navbar, and more debugged routes. On the birth and scan component pages, more fields have been added to forms, enhancing the user experience and usability of the app.